Update the graph
